\name{plots-package}
\alias{plots-package}
\alias{plots}
\docType{package}
\title{
\packageTitle{plots}
}
\description{
\packageDescription{plots}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{plots}
\packageIndices{plots}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{plots}

Maintainer: \packageMaintainer{plots}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
