% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.r
\name{plotTypes}
\alias{plotTypes}
\title{Save plots in various filetypes.}
\usage{
plotTypes(plotfx, filename, directory = "", filetype = c("pdf"),
  width = 6.5, height = 9, units = "in", res = 600,
  colormodel = "srgb")
}
\arguments{
\item{plotfx}{All the plotting commands to make the desired plot, wrapped as a function.}

\item{filename}{The desired filename of the output plot file.}

\item{directory}{If specified, the directory where the figure is to be saved.}

\item{filetype}{Filetypes to save. Possibilities include \code{pdf}, \code{png}, \code{jpeg}, \code{tiff}, \code{bmp}, and \code{eps}, or any combination therein. Defaults to \code{pdf}.}

\item{width}{Width of the figure. Defaults to \code{6.5}.}

\item{height}{Height of the figure. Defaults to \code{9}.}

\item{units}{Units on the width and height measurements (if relevant). Possibilities include \code{in}, \code{px}, \code{cm}, or \code{mm}. Defaults to \code{in}.}

\item{res}{Resolution of the figure (if relevant). Defaults to \code{600} dpi.}

\item{colormodel}{Specifies the color model (if relevant). Possibilities include (\code{srgb}, \code{gray}, or \code{cmyk}). Defaults to \code{srgb}.}
}
\value{
Creates and saves files of the plot under the desired filetypes.
}
\description{
Allows plots to be saved as a variety of filetypes, or to save the same plot under multiple filetypes.
}
\details{
The \code{filename} argument should exclude the filetype from its name (e.g., type \code{"Fig1"}, not \code{"Fig1.pdf"}).

If \code{directory} is not specified, files are saved to the working directory. If desired, filepaths can also be specified in the \code{filename} argument. For example [\code{filename = "PlotsFolder/Fig1"}] and [\code{filename = "Fig1", directory = "PlotsFolder"}] are equivalent.

Plots can be saved as more than one filetype using, e.g., \code{filetype = c("pdf", "jpeg")}.
}
\section{Warning}{

The argument \code{res} is meaningless for \code{pdf} and \code{eps} filetypes, which are vector graphics (no native resolution). Similarly, using \code{units = "px"} is meaningless for \code{pdf} and \code{eps} filetypes; if pixel units are set, the output for these filetypes will be in units of inches. Finally, \code{colormodel} only applies to \code{pdf} and \code{eps} formats; other filetypes lack this functionality.
}
\examples{
## Wrap the desired plot commands as a function
foo<-function(){plot(0,0,main='Hiya!')}

## Save the plot to the working directory with the filename "HelloWorld" as a pdf and png file 
plotTypes(foo,'HelloWorld',filetype=c('pdf','png'))
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}
}
\seealso{
\code{\link{pdf}}, \code{\link{png}}, \code{\link{jpeg}}, \code{\link{tiff}}, \code{\link{bmp}}, and \code{\link{setEPS}}, which this function wraps.
}
\concept{
plot, plotting, save
}

